Move some things to mage/public/bindings
