+++ examples/mage/mage_child.cc
@@ -10,8 +10,8 @@
 #include "base/scheduling/scheduling_handles.h"
 #include "base/scheduling/task_loop.h"
 #include "base/threading/thread.h"
-#include "mage/public/receiver.h"
-#include "mage/public/remote.h"
+#include "mage/public/bindings/receiver.h"
+#include "mage/public/bindings/remote.h"
 #include "mage/public/core.h"
 
 #include "examples/mage/magen/child_process.magen.h"  // Generated.
