+++ mage/demo/parent.cc
@@ -7,9 +7,9 @@
 #include <memory>
 
 #include "base/scheduling/task_loop_for_io.h"
-#include "mage/public/remote.h"
+#include "mage/public/bindings/remote.h"
 #include "mage/public/core.h"
-#include "mage/public/handles.h"
+#include "mage/public/bindings/message_pipe.h"
 #include "mage/demo/magen/demo.magen.h" // Generated.
 
 int main() {
