+++ mage/public/bindings/message_pipe.h
