+++ mage/public/bindings/receiver.h
@@ -5,7 +5,7 @@
 
 #include "base/scheduling/scheduling_handles.h"
 #include "base/threading/thread_checker.h"
-#include "mage/public/handles.h"
+#include "mage/public/bindings/message_pipe.h"
 #include "mage/public/message.h"
 
 namespace mage {
