+++ mage/public/bindings/receiver_delegate.h
