+++ mage/public/parser/classes.tmpl
@@ -5,9 +5,9 @@
 #include <vector>
 
 #include "base/check.h"
-#include "mage/public/handles.h"
+#include "mage/public/bindings/message_pipe.h"
+#include "mage/public/bindings/receiver_delegate.h"
 #include "mage/public/message.h"
-#include "mage/public/receiver_delegate.h"
 
 namespace magen {
