+++ mage/test/magen/BUILD
@@ -1,5 +1,5 @@
 load("@rules_cc//cc:defs.bzl", "cc_library")
-load("//mage/parser:magen_idl.bzl", "magen_idl")
+load("//mage/public/parser:magen_idl.bzl", "magen_idl")
 
 package(default_visibility = ["//visibility:public"])
