+++ mage/test/pass_pipe_back_and_forth.cc
@@ -10,10 +10,10 @@
 #include "base/scheduling/scheduling_handles.h"
 #include "base/scheduling/task_loop_for_io.h"
 #include "base/threading/thread_checker.h"
-#include "mage/public/receiver.h"
-#include "mage/public/remote.h"
+#include "mage/public/bindings/receiver.h"
+#include "mage/public/bindings/remote.h"
 #include "mage/public/core.h"
-#include "mage/public/handles.h"
+#include "mage/public/bindings/message_pipe.h"
 #include "mage/test/magen/handle_accepter.magen.h"  // Generated.
 
 class HandleAccepterImpl final : public magen::HandleAccepter {
